requirements.txt
bs4
lxml
pyyaml
xmltodict
requests